implemented growing() and harvest()
wrote function so that crops grow and can be harvested. did some preliminary bugtesting to make sure it worked with plant() and remove(). minor tweaks to prefabs.

also ive decided to stop branching and just work on main now. branching makes sense for a team project but not so much for personal